Converted totalCentimeters to uint64_t (Resolves #25)

+++ main/newsettings.h
@@ -435,7 +435,7 @@ public:
         ConfigWrapperLegacy<bool> allowPresetSwitch     {true,                                   DoReset,   {},                         "lockAlwPresetSw"     };
         ConfigWrapperLegacy<bool> keepLockedAfterReboot {false,                                  DoReset,   {},                         "keepLocked"          };
         ConfigWrapperLegacy<bool> locked                {false,                                  DoReset,   {},                         "currentlyLocked"     };
-        ConfigWrapperLegacy<std::optional<uint32_t>> rememberMe{std::nullopt,                    DoReset,   {},                         "rememberMe"          };
+        ConfigWrapperLegacy<std::optional<uint64_t>> rememberMe{std::nullopt,                    DoReset,   {},                         "rememberMe"          };
         std::array<ConfigWrapperLegacy<int8_t>, 4> pin {
             ConfigWrapperLegacy<int8_t>                 {1,                                      DoReset,   MinMaxValue<int8_t, 0, 9>,  "lockscreenPin0"      },
             ConfigWrapperLegacy<int8_t>                 {2,                                      DoReset,   MinMaxValue<int8_t, 0, 9>,  "lockscreenPin1"      },
@@ -451,7 +451,7 @@ public:
     } lockscreen;
 
     struct {
-        ConfigWrapperLegacy<uint32_t> totalCentimeters  {0,                                      DoReset,   {},                         "totalCentimeter"     };
+        ConfigWrapperLegacy<uint64_t> totalCentimeters  {0,                                      DoReset,   {},                         "totalCentimeter"     };
     } savedStatistics;
 
     struct {
